autonomous-knowledge-graph-platform/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
neo4j==5.22.0
openai==1.35.10
google-generativeai==0.7.2
pandas==2.2.2
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
email-validator==2.2.0
pytest==8.2.2
